BackTask/requirements.txt
Django==2.2.3
Pillow==6.1.0
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
djangorestframework
djangorestframework-simplejwt
